% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDGEList.R
\name{addCellTypeProportionsPCA}
\alias{addCellTypeProportionsPCA}
\title{Parse the cell type proportions PCA file and add to the metrics DF.}
\usage{
addCellTypeProportionsPCA(
  cellTypeProportionsPCAFile,
  metricsDF,
  numPCs = 4,
  scalePCs = FALSE
)
}
\arguments{
\item{cellTypeProportionsPCAFile}{Path to the PCA file containing cell type proportions.}

\item{metricsDF}{A data frame containing cell-level metadata, which will be updated with PCA scores.}

\item{numPCs}{The number of principal components to retain from the PCA file (default is 4).}

\item{scalePCs}{Logical; if \code{TRUE}, scales the PCA scores to have mean 0 and standard deviation 1 (default is \code{FALSE}).}
}
\value{
A data frame identical to \code{metricsDF} but with additional columns for the PCA scores of cell type proportions.
}
\description{
This function reads a PCA file containing cell type proportions,
scales the PCA scores if requested, and adds them to the metrics data frame.
}
