% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_summary_barplot.R
\name{annotate_de_chromosome_group}
\alias{annotate_de_chromosome_group}
\title{Annotate differential-expression results by chromosome group}
\usage{
annotate_de_chromosome_group(d, contig_yaml_file, reduced_gtf_file)
}
\arguments{
\item{d}{A data frame or data.table containing a \code{gene} column.}

\item{contig_yaml_file}{Path to a YAML file mapping contigs to classes.}

\item{reduced_gtf_file}{Path to the reduced GTF-like annotation file.}
}
\value{
A data.table with an added \code{chromosome_group} column.
}
\description{
Annotate differential-expression results by chromosome group
}
