% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDGEList.R
\name{build_eQTL_covariates}
\alias{build_eQTL_covariates}
\title{Build eQTL covariates for tensorQTL}
\usage{
build_eQTL_covariates(
  manifest_file,
  metacell_dir,
  cell_metadata_file,
  metadata_columns,
  force_category_columns = c(),
  outDir
)
}
\arguments{
\item{manifest_file}{Path to the manifest file, with required columns: \code{output_name}, \code{cell_type_name}, \code{region_name}, and \code{cell_type_filter_str}.}

\item{metacell_dir}{Path to the directory containing \code{*.metacells.txt.gz} files.}

\item{cell_metadata_file}{Path to the TSV file containing donor-level metadata.}

\item{metadata_columns}{Character vector of metadata column names to include from \code{cell_metadata_file}.  The
program will automatically capture donor id.}

\item{force_category_columns}{A vector of column names in \code{metadata_columns} that should be treated as categorical variables}

\item{outDir}{Directory to save the covariates files.  Each file will have the matching prefix of the metacell file.}
}
\description{
This is functionally very similar to build_merged_dge, but the eQTL data keeps each
cell type / region combination as a separate input matrix, and needs a matching covariate file.
The covariates are additionally one-hot encoded in preparation for tensorQTL.
}
\details{
The output covariates file for each cell type / region combination is in the expected format for PrepareEqtlCovariates:
In the first row/column "id,".  The header column contains donor IDs, each row's first column is the covariate name, followed by values.
}
