% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDGEList.R
\name{compareDGEList}
\alias{compareDGEList}
\title{Compare two DGEList objects for near equality}
\usage{
compareDGEList(dge1, dge2, tol = 1e-08)
}
\arguments{
\item{dge1}{A DGEList object}

\item{dge2}{A DGEList object (e.g., loaded)}

\item{tol}{Numeric tolerance for comparing numeric values (default 1e-8)}
}
\value{
TRUE if equal within tolerance, otherwise throws an error describing the mismatch
}
\description{
Compare two DGEList objects for near equality
}
