% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDGEList.R
\name{createMetadataForAggregationWorkflow}
\alias{createMetadataForAggregationWorkflow}
\title{Reads data from an ScRnaAggregationWorkflow metacell file, and creates the corresponding metadata stub to be filled in
This assumes that the metacell column names encode the donor and cell type with a colon separator, e.g. "MD1234:Astrocyte".
The metadata will be created with columns for donor and cell type,
which can then be filled in with additional metadata from the manifest or other sources.}
\usage{
createMetadataForAggregationWorkflow(
  metacell_file_list,
  outDir,
  outName = "donor_rxn_DGEList",
  validate_round_trip = T
)
}
\arguments{
\item{metacell_file_list}{A list of paths to metacell files to be processed.}

\item{outDir}{Directory where the merged DGEList object will be saved as two gzipped TSV files.}

\item{outName}{Prefix for the output files.}

\item{validate_round_trip}{Logical; if TRUE, validates the saved DGEList by loading it back and comparing to the original.}
}
\description{
Reads data from an ScRnaAggregationWorkflow metacell file, and creates the corresponding metadata stub to be filled in
This assumes that the metacell column names encode the donor and cell type with a colon separator, e.g. "MD1234:Astrocyte".
The metadata will be created with columns for donor and cell type,
which can then be filled in with additional metadata from the manifest or other sources.
}
