% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Run differential expression analysis for each cell type in the DGEList.}
\usage{
differential_expression(
  data_dir,
  data_name,
  randVars,
  fixedVars,
  contrast_file,
  outPDF,
  result_dir
)
}
\arguments{
\item{data_dir}{Directory containing the DGEList data.}

\item{data_name}{Name of the DGEList data file (without extension).}

\item{randVars}{Vector of random effect variables.}

\item{fixedVars}{Vector of fixed effect variables.}

\item{contrast_file}{Path to the file containing contrast definitions.}

\item{outPDF}{Optional path to output PDF file for plots.}

\item{result_dir}{Directory to save the differential expression results.}
}
\description{
Run differential expression analysis for each cell type in the DGEList.
}
