% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDGEList.R
\name{loadDGEList}
\alias{loadDGEList}
\title{Load a DGEList from disk saved by saveDGEList with a filename prefix}
\usage{
loadDGEList(dir, prefix = "dge")
}
\arguments{
\item{dir}{Directory path containing the files}

\item{prefix}{Character prefix used when saving (default "dge").}
}
\value{
A DGEList object
}
\description{
Load a DGEList from disk saved by saveDGEList with a filename prefix
}
