% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_summary_barplot.R
\name{make_sex_de_chromosome_counts}
\alias{make_sex_de_chromosome_counts}
\title{Count significant genes by chromosome group}
\usage{
make_sex_de_chromosome_counts(d)
}
\arguments{
\item{d}{Annotated differential-expression results.}
}
\value{
A data.table with gene counts by cell type, interaction, and
  chromosome group.
}
\description{
Count significant genes by chromosome group
}
