% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_vs_pmi_rqs.R
\name{plot_nested_models}
\alias{plot_nested_models}
\title{Compare two nested models with and without the tested trait vs age}
\usage{
plot_nested_models(
  tested_trait = "RQS",
  in_dir_full_model,
  in_dir_partial_model,
  cellTypeListFile = NULL,
  output_file
)
}
\arguments{
\item{tested_trait}{The covariate added in the full model}

\item{in_dir_full_model}{The directory containing the full model results}

\item{in_dir_partial_model}{The directory containing the partial results}

\item{cellTypeListFile}{A file containing a list of cell types to generate plots for.
If null generate plots for all cell types.}

\item{output_file}{A PDF report with summary plots and per cell type / region plots of the effect sizes.}
}
\description{
Does adding the tested trait change the age effect size?
}
\details{
The setup for this test is important - it's critical to have the exact same
set of data points in both model runs, and only change the model by adding the
tested trait to the full model.
}
