% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_summary_barplot.R
\name{plot_sex_de_chromosome_counts}
\alias{plot_sex_de_chromosome_counts}
\title{Plot significant-gene counts by chromosome group}
\usage{
plot_sex_de_chromosome_counts(counts, title, alpha = 0.05)
}
\arguments{
\item{counts}{Output from \code{make_sex_de_chromosome_counts()}.}

\item{title}{Plot title.}

\item{alpha}{Adjusted P-value threshold used to select genes.}
}
\value{
A ggplot object.
}
\description{
Plot significant-gene counts by chromosome group
}
