% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variancePartition.R
\name{prepare_data_for_differential_expression}
\alias{prepare_data_for_differential_expression}
\title{Prepare data for variance partition or differential expression}
\usage{
prepare_data_for_differential_expression(
  data_dir,
  data_name,
  randVars,
  fixedVars,
  interaction_var = NULL,
  force_as_factor = c("imputed_sex")
)
}
\arguments{
\item{data_dir}{Character. Directory containing the precomputed DGEList.}

\item{data_name}{Character. Prefix or name used to load the DGEList object.}

\item{randVars}{Character vector. Names of random metadata variables for MDS coloring.}

\item{fixedVars}{Character vector. Names of fixed metadata variables for MDS grouping.}

\item{interaction_var}{Character. Optional. Name of variable to include as an interaction term
if running differential expression.  Optional, not used for VariancePartition.}

\item{force_as_factor}{Character vector. Optional. Column names to force as factors.  Important when
there are encodings of categorical variables that use numeric values.  Default is c("imputed_sex") which
is useful for the BICAN data, but this can be altered as needed.}
}
\value{
A list containing the DGEList object, fixed variables, and random variables.
}
\description{
Load a precomputed DGEList, validate sample variables, scale PCs,
filter samples by library size and runs variance partition both on the full
data set and by cell type.
}
\details{
A special field in the DGEList samples metadata is num_nuclei, which is the total number of
nuclei captured in the single cell assay. If present, this is used in the variance partition model
to control for technical variation.  The value is converted to a Z-score of log10(num_nuclei) and
added to the fixed effects.
}
