% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashR.R
\name{remove_common_tokens}
\alias{remove_common_tokens}
\title{Remove common tokens from strings}
\usage{
remove_common_tokens(x, delim = "_")
}
\description{
Useful to shorten condition names by removing shared prefixes/suffixes
}
