% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentvariabileinclusion.R
\name{restrictDGEListToSamples}
\alias{restrictDGEListToSamples}
\title{Restrict a DGEList to a random subset of samples}
\usage{
restrictDGEListToSamples(dge, num_samples)
}
\arguments{
\item{dge}{A DGEList object containing sample metadata}

\item{num_samples}{Integer. Number of samples to randomly select; if NULL, all samples are retained.}
}
\value{
A DGEList object containing only the selected samples, with library sizes retained.
}
\description{
Restrict a DGEList to a random subset of samples
}
