% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variancePartition.R
\name{runVariancePartition}
\alias{runVariancePartition}
\title{Run variance partition and generate QC plots}
\usage{
runVariancePartition(
  data_dir,
  data_name,
  randVars,
  fixedVars,
  outPDF,
  variance_partition_result_dir
)
}
\arguments{
\item{data_dir}{Character. Directory containing the precomputed DGEList.}

\item{data_name}{Character. Prefix or name used to load the DGEList object.}

\item{randVars}{Character vector. Names of random metadata variables for MDS coloring.}

\item{fixedVars}{Character vector. Names of fixed metadata variables for MDS grouping.}

\item{outPDF}{QC and variance partition plots are saved to this single PDF file.}

\item{variance_partition_result_dir}{Character. Directory to save variance partition RDS objects to inspect later.}
}
\description{
Load a precomputed DGEList, validate sample variables, scale PCs,
filter samples by library size and runs variance partition both on the full
data set and by cell type.
}
\details{
A special field in the DGEList samples metadata is num_nuclei, which is the total number of
nuclei captured in the single cell assay. If present, this is used in the variance partition model
to control for technical variation.  The value is converted to a Z-score of log10(num_nuclei) and
added to the fixed effects.
}
