% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDGEList.R
\name{saveDGEList}
\alias{saveDGEList}
\title{Save a DGEList to disk as two gzipped TSV files with a filename prefix}
\usage{
saveDGEList(dge, dir, prefix = "dge")
}
\arguments{
\item{dge}{A DGEList object from edgeR}

\item{dir}{Directory path where the files will be written (will be created if needed)}

\item{prefix}{Character prefix for the filenames (default "dge").
Will produce files: prefix_counts.tsv.gz and prefix_samples.tsv.gz}
}
\description{
Save a DGEList to disk as two gzipped TSV files with a filename prefix
}
