% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentvariabileinclusion.R
\name{scale_PC_cols}
\alias{scale_PC_cols}
\title{Scale all PC columns in a data.frame to unit variance}
\usage{
scale_PC_cols(df)
}
\arguments{
\item{df}{A data.frame containing PC columns (named "PC1", "PC2", etc.)}
}
\value{
The original data.frame with PC columns scaled to unit variance
}
\description{
Scale all PC columns in a data.frame to unit variance
}
