% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentvariabileinclusion.R
\name{validateSampleVars}
\alias{validateSampleVars}
\title{Validate that required sample variables are present}
\usage{
validateSampleVars(dge, required_vars)
}
\arguments{
\item{dge}{A DGEList object}

\item{required_vars}{Character vector of variable names that must exist in dge$samples}
}
\value{
TRUE if all are present, otherwise throws an error listing missing variables
}
\description{
Validate that required sample variables are present
}
