% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_chromosome_leakage.R
\name{write_sex_chromosome_leakage}
\alias{write_sex_chromosome_leakage}
\title{Write sex-chromosome leakage results to a tab-delimited file}
\usage{
write_sex_chromosome_leakage(df, file)
}
\arguments{
\item{df}{A data frame produced by
\code{\link{compute_sex_chromosome_leakage}}.}

\item{file}{Output file path.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Write sex-chromosome leakage results to a tab-delimited file
}
\keyword{internal}
