% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDGEList.R
\name{addMixedAssayType}
\alias{addMixedAssayType}
\title{Flag donors with multiple assay types in the same region}
\usage{
addMixedAssayType(metricsDF)
}
\arguments{
\item{metricsDF}{A data frame containing at least the columns
 `single_cell_assay`, `donor_external_id`, and `brain_region_abbreviation`.
Each row represents a nuclei for a donor with a corresponding assay type / region}
}
\value{
A data frame identical to `metricsDF` except that for donors associated
  with more than one distinct `single_cell_assay` in a region, the `single_cell_assay`
  value is set to `"mixed"` in all rows for that donor-region combination.
}
\description{
This function identifies donors that are associated with more than one unique
sequencing assay in the same brain region and sets their `sequencing_assay`
value to `"NextGEM:GEMX"` in the input data frame.
}
