% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Run differential expression analysis for each cell type in the DGEList.}
\usage{
differential_expression(
  data_dir,
  data_name,
  randVars,
  fixedVars,
  contrast_file,
  interaction_var = NULL,
  cellTypeListFile = NULL,
  outPDF = NULL,
  result_dir
)
}
\arguments{
\item{data_dir}{Directory containing the DGEList data.}

\item{data_name}{Name of the DGEList data file (without extension).}

\item{randVars}{Vector of random effect variables.}

\item{fixedVars}{Vector of fixed effect variables.}

\item{contrast_file}{Path to the file containing contrast definitions.}

\item{interaction_var}{Optional name of a variable to test for interactions with the contrast variable. If NULL, no interaction terms are added.}

\item{cellTypeListFile}{A file containing an explicit list of cell types to test.  If NULL, all cell types in the DGEList will be tested.}

\item{outPDF}{Optional path to output PDF file for plots.}

\item{result_dir}{Directory to save the differential expression results.}
}
\description{
This uses a means model (~0 + fixedVars) and a random effects model for the specified random variables.
For each contrast group, the fixed effects are reordered so the contrast group is first, which
makes all levels of the contrast available for comparison.
}
